python-pipeline/test_key.py
# test_key.py
from google import genai
from dotenv import load_dotenv
import os
load_dotenv()
client = genai.Client(api_key=os.getenv("GEMINI_API_KEY"))
r = client.models.generate_content(model="gemini-2.0-flash", contents="Say hello")
print(r.text)