python-pipeline/requirements.txt
google-genai
pinecone
mysql-connector-python
python-dotenv
 